+++ game-backend/src/main/java/com/becky/Gun.java
@@ -1,5 +1,7 @@
 package com.becky;
 
+import com.becky.world.entity.Player;
+
 /**
  * A gun fires bullets by creating them. Is simple.
  */
