+++ game-backend/src/main/java/com/becky/networked/message/ClientInputStateUpdate.java
@@ -1,4 +1,4 @@
-package com.becky.networked;
+package com.becky.networked.message;
 
 import org.json.JSONObject;
