+++ game-backend/src/main/java/com/becky/world/entity/Bullet.java
@@ -1,4 +1,4 @@
-package com.becky;
+package com.becky.world.entity;
 
 import java.awt.geom.Point2D;
